feat: move out search for grid to separate component to avoid spaghetti code
